Refactor authentication context structure and update imports across components
